Feat: planning
Added a few files to scaffold a prototype of intended
behavior. Manifest files that govern how code is
generated are placed in the manifests directory. The
app to be generated will use a json file in the current
directory and is used to specify which manifests to
source, database type, which models to create, and
which pages will be created on the basis of declared
models.